SignUpPage.js to tsx
